add some test for NodesSpec and BundlebridgeSpec
